Made instance variables private
Previous code had public getters and setters, and no access modifier
for instance variables. With no access modifier, the instance variables
would be visible to that particular class, and other classes in the
package. The instance variables would not be visible to subclasses and
classes in other packages.
Since there are public getters and setters for all of the instance
variables, it would make sense to make the instance variables be
private.